it's works!
it's works!

+++ Photo.Print/Layout/Base/Area.cs
@@ -8,14 +8,14 @@ namespace Photo.Print.Layout.Base
 {
     public struct Area
     {
-        public float Left { get; set; }
-        public float Top { get; set; }
+        public double Left { get; set; }
+        public double Top { get; set; }
 
-        public float Width { get; set; }
-        public float Height { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
 
         public Photo? Photo { get; set; }
-        public int Angle { get; set; }
+        public float Angle { get; set; }
         public FitType FitType { get; set; }
     }
 }
